feat: Connect with database

+++ src/main/java/com/choikang/poor/the_poor_back/dto/RankingDTO.java
@@ -11,7 +11,7 @@ import lombok.*;
 @NoArgsConstructor
 public class RankingDTO {
     private int rankingNum;
-    private int rankingUserNum;
+    private long rankingUserNum;
     private String rankingUserName;
     private int rankingScore;
     private int userTotalScore;
